PROJECT RESTART
So to add a little bit of context behind this decision
I've struggled to get going with this task as I
found it really overwhelming trying to decide which
function should go to which model, how to implement
and import them. Causing anxiety and stress whenever I've
tried to code something, doubting myself etc. So I've decided
to "start over" code it in "one" file to just get going
and later think maybe think about separating each module.